Misc: fixed module inception error in clippy
